# openapi.yaml
openapi: 3.0.3
info:
  title: Armada Convoy API
  description: REST API for the Armada convoy and social app
  version: 1.0.0
  contact:
    name: Armada Team
    email: support@armada.app
servers:
  - url: https://convoy-api.onrender.com
    description: Production server
  - url: http://localhost:8080
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /me:
    get:
      summary: Get current user
      description: Get the current authenticated user's profile
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      summary: Update current user
      description: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                  pattern: '^[a-zA-Z0-9_]+$'
                bio:
                  type: string
                  maxLength: 500
                avatarUrl:
                  type: string
                  format: uri
                settings:
                  $ref: '#/components/schemas/UserSettings'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/search:
    get:
      summary: Search users
      description: Search for users by username or email
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: limit
          in: query
          description: Number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /users/{id}:
    get:
      summary: Get user by ID
      description: Get user profile by ID
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/by-username/{username}:
    get:
      summary: Get user by username
      description: Get user profile by username
      parameters:
        - name: username
          in: path
          required: true
          description: Username
          schema:
            type: string
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: johndoe
        avatarUrl:
          type: string
          format: uri
          example: https://example.com/avatar.jpg
        bio:
          type: string
          example: Adventure enthusiast and convoy leader
        settings:
          $ref: '#/components/schemas/UserSettings'
        stats:
          $ref: '#/components/schemas/UserStats'
        friends:
          type: array
          items:
            $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserSettings:
      type: object
      properties:
        showConvoys:
          type: string
          enum: [everyone, friends, private]
          example: friends
        showStats:
          type: string
          enum: [everyone, friends, private]
          example: friends
        showProfile:
          type: string
          enum: [everyone, friends, private]
          example: everyone
        notifications:
          type: object
          properties:
            convoyInvites:
              type: boolean
              example: true
            friendRequests:
              type: boolean
              example: true
            messages:
              type: boolean
              example: true
            achievements:
              type: boolean
              example: true
    UserStats:
      type: object
      properties:
        convoysCreated:
          type: integer
          example: 12
        convoysJoined:
          type: integer
          example: 45
        totalMiles:
          type: number
          example: 1250.5
        friendsCount:
          type: integer
          example: 23
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
        total:
          type: integer
          example: 100
        hasMore:
          type: boolean
          example: true
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...